// commentSold/commentsoldfront/src/constant/enums.ts
export enum E_USER_TYPE {
  Influencer = "influencer",
  Brand = "brand",
  BrandInfluencer = "brand-influencer",
}

export enum GENDER {
  Male = "MALE",
  Female = "FEMALE",
  Other = "OTHER",
}

export enum USER_STATUS {
  Active = "ACTIVE",
  InActive = "INACTIVE",
  Suspended = "SUSPENDED",
}

export enum E_PLATFORMS {
  Facebook = "facebook",
  Instagram = "instagram",
  Youtube = "youtube",
  TikTok = "tiktok",
  Linkedin = "linkedin",
}